Init WinConfigurator DBAccess

+++ COREMODEL/USERS.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -14,11 +15,11 @@ namespace COREMODEL
         public String USERNAME { get; set; }
         public String PASSWD { get; set; }
 
-        public Int64 ID { get { return _id; } }
+        public Int32 ID { get { return _id; } }
 
         public State STATE { get { return _state; } }
 
         private State _state { get; set; }
-        private Int64 _id { get; set; }
+        private Int32 _id { get; set; }
     }
 }
